+++ water-be/internal/tools/process_windows.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package tools
+
+import "os/exec"
+
+func configureCommandProcessGroup(cmd *exec.Cmd) {}
+
+func terminateCommandProcessGroup(cmd *exec.Cmd) error {
+	if cmd.Process == nil {
+		return nil
+	}
+	return cmd.Process.Kill()
+}